# GrooverAdm/ClientApp/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Groover API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:44335"
    }
  ],
  "paths": {
...
    "/api/place/list": {
      "get": {
        "tags": [
          "Place"
        ],
        "summary": "Retrieves establishments",
        "operationId": "Place_GetEstablishmentsAll",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            },
            "x-position": 2
          },
          {
            "name": "mine",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "x-position": 3
          },
          {
            "name": "pendingReview",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place": {
      "get": {
        "tags": [
          "Place"
        ],
        "summary": "Retrieves the establishments surrounding [, ] in \nthe distance provided",
        "operationId": "Place_GetEstablishments",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "Latitude",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 1
          },
          {
            "name": "lon",
            "in": "query",
            "description": "Longitude",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 2
          },
          {
            "name": "distance",
            "in": "query",
            "description": "Distance in METERS",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 3
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 4
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Place"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Place"
        ],
        "summary": "Creates an establishment",
        "operationId": "Place_CreateEstablishment",
        "requestBody": {
          "x-name": "establishment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Place"
        ],
        "summary": "Updates a establishment",
        "operationId": "Place_UpdateEstablishment",
        "requestBody": {
          "x-name": "establishment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Place"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Place"
        ],
        "summary": "Deletes an establishment",
        "operationId": "Place_DeleteEstablishment",
        "parameters": [
          {
            "name": "establishmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place/recommended": {
      "get": {
        "tags": [
          "Place"
        ],
        "summary": "Obtiene los lugares recomendados en funcion de la playlist enviada",
        "operationId": "Place_GetRecommendedEstablishmentsForPlaylist",
        "parameters": [
          {
            "name": "playlistId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 2
          },
          {
            "name": "lon",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 3
          },
          {
            "name": "distance",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 4
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 5
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            },
            "x-position": 6
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComparedPlace"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place/recommended/top": {
      "get": {
        "tags": [
          "Place"
        ],
        "summary": "Obtiene los lugares recomendados en funcion del top50 del usuario",
        "operationId": "Place_GetRecommendedEstablishmentsForTop",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 1
          },
          {
            "name": "lon",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 2
          },
          {
            "name": "distance",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 3
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 4
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            },
            "x-position": 5
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComparedPlace"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place/{id}": {
      "get": {
        "tags": [
          "Place"
        ],
        "summary": "Gets a single place from the db",
        "operationId": "Place_GetPlace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place/{establishmentId}/song": {
      "post": {
        "tags": [
          "Place"
        ],
        "summary": "We gucci",
        "operationId": "Place_RecognizeSong",
        "parameters": [
          {
            "name": "establishmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "song",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Song"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place/{placeId}/rate": {
      "post": {
        "tags": [
          "Place"
        ],
        "summary": "Rates a place for a user, if the register is already existing, it updates it",
        "operationId": "Place_RatePlace",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "value",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/place/{placeId}/review": {
      "post": {
        "tags": [
          "Place"
        ],
        "summary": "Rates a place for a user, if the register is already existing, it updates it",
        "operationId": "Place_ReviewPlace",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "review",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceReview"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Place"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Place": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Geolocation"
              }
            ]
          },
          "mainPlaylist": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Playlist"
              }
            ]
          },
          "weeklyPlaylists": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {
              "$ref": "#/components/schemas/Playlist"
            }
          },
          "ratings": {
            "type": "number",
            "format": "double"
          },
          "ratingCount": {
            "type": "integer",
            "format": "int32"
          },
          "owners": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ListableUser"
            }
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "geohash": {
            "type": "string",
            "nullable": true
          },
          "recognizedMusic": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/RecognizedSong"
            }
          },
          "timetables": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Timetable"
            }
          },
          "approved": {
            "type": "boolean"
          },
          "pendingReview": {
            "type": "boolean"
          },
          "reviewComment": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Geolocation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Playlist": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "songs": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "snapshotVersion": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "type": "string",
            "nullable": true
          },
          "changed": {
            "type": "boolean"
          }
        }
      },
      "Song": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "artists": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Artist"
            }
          }
        }
      },
      "Artist": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ListableUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "born": {
            "type": "integer",
            "format": "int32"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "admin": {
            "type": "boolean"
          }
        }
      },
      "RecognizedSong": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Song"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "count": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "Timetable": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schedules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Schedule"
            }
          },
          "day": {
            "$ref": "#/components/schemas/DayOfWeek"
          }
        }
      },
      "Schedule": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DayOfWeek": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Sunday",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ]
      },
      "ComparedPlace": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Place"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "similitude": {
                "type": "number",
                "format": "double"
              }
            }
          }
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
      "PlaceReview": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "approved": {
            "type": "boolean"
          },
          "reviewComment": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Type into the textbox: Bearer {your JWT token}.",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}